requirements.txt
azure-cosmos
azure-identity
pydantic
pydantic-settings
opentelemetry-api
opentelemetry-sdk
azure-monitor-opentelemetry
